Refactor Psychiatric Assessment and Physical Examination forms
