+++ src/components/forms/fields/CheckboxForm.js
@@ -8,7 +8,7 @@ import {
 } from "react-native";
 import { useFormikContext } from "formik";
 import PropTypes from "prop-types";
-import FormError from "../FormError";
+import FormError from "./FormError";
 import { Colors, Typography } from "../../../constants/styles";
 import FontForgeIcon from "../../common/FontForgeIcon";
