+++ src/constants/values/initalFormValues.js
@@ -1,0 +1,196 @@
+export const initialPatient = {
+  name: "",
+  surname: "",
+  sex: "male",
+  pesel: "",
+  date_of_birth: "",
+  weight: "0",
+  bmi: "0",
+  height: "0",
+  note: "",
+  phone: "",
+  person_guard: "",
+  phone_guard: "",
+};
+
+export const initialBasicData = {
+  reason_of_report: "",
+  major_ailments: "",
+  suicidal_thoughts_choice: null,
+  suicidal_thoughts: "",
+  other_ailments: "",
+  past_diseases_choice: "",
+  past_diseases: "",
+  past_psychiatric_treatment: false,
+  first_hospitalization: "",
+  hospitalization_times: 0,
+  pharmacotherapy: "",
+  psychotherapy: "",
+  family_therapy: "",
+  medications: "",
+  allergies: "",
+  hygiene: "",
+  education_choice: "",
+  education: "",
+  professional_status: "",
+  social_conditions: "",
+  social_assistance_choice: null,
+  social_assistance: "",
+  social_level_changes: "",
+  development_data: "",
+  family_situation: "",
+  family_situation_changes: "",
+  family_relationships: "",
+  hereditary_taint: "",
+  physical_activity: "",
+  self_mutilation: "",
+  occupational_exposure: "",
+  alcohol: "",
+  nicotine: "",
+  psychoactive_substances: "",
+  diet_choice: "",
+  diet: "",
+  family_interview: "",
+};
+
+export const initialPhysicalExamination = {
+  general_conditions: "dobry",
+  blood_pressure: "",
+  pulse: "",
+  body_temperature: "",
+  body_build_type: "normosteniczna",
+  skin_appearance: "",
+  skin_colour: "",
+  skin_humidity: "",
+  skin_swelling: "",
+  skin_scars: "",
+  lymphatic_gland_examined: false,
+  head_appearance_choice: false,
+  head_appearance: "",
+  head_percussion_tenderness: "",
+  head_eyeballs: "",
+  ears: "",
+  nose: "",
+  mouth_teeth: "uporządkowane",
+  mucous_membrane_choice: "",
+  mucous_membrane: "",
+  neck_throat_tonsil: "",
+  neck_appearance: "",
+  neck_thyroid_choice: "",
+  neck_thyroid: "",
+  chest_choice: "",
+  chest: "",
+  breath_frequency: "",
+  chest_percussion_choice: "",
+  chest_percussion: "",
+  chest_auscultation_choice: false,
+  chest_auscultation: "",
+  cardiovascular_appearance: "",
+  cardiovascular_efficient: "",
+  cardiovascular_auscultation: "",
+  cardiovascular_pulse_choice: null,
+  cardiovascular_pulse: "",
+  stomach: "wysklepiony w poziomie klp",
+  stomach_hernia: "",
+  stomach_painless: "",
+  stomach_auscultation: "",
+  stomach_percussion: "",
+  stomach_physical_examination: "",
+  stomach_symptoms: "",
+  legs_swelling: "",
+  legs_veins: "",
+  locomotor_joint_outline: true,
+  locomotor_limb_mobility: true,
+  muscle_strength_tension: true,
+  nervous_meningeal_signs: "",
+  nervous_focal_damage: "",
+};
+
+export const initialPsychiatricAssessment = {
+  general_appearance: "",
+  outfit_choice: "",
+  appearance_care: "",
+  unusual_features_gestures: "",
+  agree_examination_choice: null,
+  agree_examination: "",
+  trusting_choice: null,
+  trusting: "",
+  aggressive_choice: null,
+  aggressive: "",
+  distance_shorten_choice: null,
+  distance_shorten: "",
+  sexualizing_contact_choice: null,
+  sexualizing_contact: "",
+  irritated_easily_choice: null,
+  irritated_easily: "",
+  fear_cause_choice: null,
+  fear_cause: "",
+  irritate_choice: null,
+  irritate: "",
+  autopsychic_orientation_choice: null,
+  autopsychic_orientation: "",
+  allopsychic_orientation_choice: null,
+  allopsychic_orientation: "",
+  answer_questions_choice: "",
+  answer_questions: "",
+  accent_difficulties: null,
+  accent: "",
+  dialect_difficulties: null,
+  dialect: "",
+  speech_speed_difficulties: null,
+  speech_speed: "",
+  speech_tone_difficulties: null,
+  speech_tone: "",
+  speech_impairment_difficulties: null,
+  speech_impairment: "",
+  aphasia_features_difficulties: null,
+  aphasia_features: "",
+  moving_way_choice: "",
+  moving_way: "",
+  attention_disorders_choice: "",
+  attention_disorders: "",
+  memory_impairment_choice: "",
+  memory_impairment: "",
+  mood_choice: "",
+  mood: "",
+  psychomotor_drive_choice: "",
+  psychomotor_drive: "",
+  affect_choice: "",
+  affect: "",
+  anxiety_choice: "",
+  anxiety: "",
+  delusions_choice: "",
+  delusions: "",
+  hallucinations_choice: "",
+  hallucinations: "",
+  illusions_choice: "",
+  illusions: "",
+  perception_disorders_choice: "",
+  perception_disorders: "",
+  abnormal_thinking_choice: "",
+  abnormal_thinking: "",
+  criticism_disturbance_choice: "",
+  criticism_disturbance: "",
+  complex_activity_choice: "",
+  complex_activity: "",
+  catatonic_parakinesis_choice: "",
+  catatonic_parakinesis: "",
+  memorizing_difficulties: null,
+  memorizing: "",
+  attention_counting_difficulties: null,
+  attention_counting: "",
+  reminding_difficulties: null,
+  reminding: "",
+  language_skills_difficulties: null,
+  language_skills: "",
+  resignation_thoughts_choice: null,
+  resignation_thoughts: "",
+  suicide_thoughts_choice: null,
+  suicide_thoughts: "",
+  suicidal_plans_choice: null,
+  suicidal_plans: "",
+  murder_fantasies_choice: null,
+  murder_fantasies: "",
+  murder_plans_choice: null,
+  murder_plans: "",
+};
